Initial Commit
Looking into demoing out what the limits are of NBTify, in regards to large NBT files with a heck ton of keys and such. Seems to crash Dovetail and WebNBT, but not Misode's extension for VSCode. This runs (fairly) nicely in Node.js and the browser, I think the other ones crash because of trying to render that content out in the DOM. Misode's probably uses proper dynamic DOM modifications, while Dovetail definitely isn't very smart yet on dynamically rendering things out. Could use much more improvements on my part :)